Server/requirements.txt
django
django-cors-headers
djangorestframework
moviepy
Pillow
djangorestframework-simplejwt
django_filter
